+++ docs/ReleaseNotes.md
@@ -1,6 +1,5 @@
 // Before any commit, use the command line: tsc --noEmit
 
-# 0.1.2.26 Translations Refactor
+# 0.1.2.27 Alert Refactor
 
-- all translations completely redesign.
-- added all missing services and components translations.
+- all functions now utilizes useCallback hook.